This commit adds a validator which returns appropriate validation functions based on the state of AKO. In case of a leader AKO, the validator returns a function that does the actual validations and in follower AKO, it returns a function which logs a message and doesn't do any validations.
