+++ internal/k8s/controller.go
@@ -1329,3 +1329,7 @@ func (c *AviController) Run(stopCh <-chan struct{}) error {
 
 	return nil
 }
+
+func (c *AviController) GetValidator() Validator {
+	return NewValidator()
+}
