+++ internal/k8s/leader_election_callbacks.go
@@ -39,6 +39,4 @@ func (c *AviController) OnStoppedLeading() {
 	lib.AKOControlConfig().SetIsLeaderFlag(false)
 	utils.AviLog.Debugf("AKO lost the leadership")
 	lib.AKOControlConfig().PodEventf(v1.EventTypeNormal, "LeaderElection", "AKO lost the leadership")
-	lib.SetDisableSync(true)
-	c.DisableSync = true
 }
